Rename package "nl.hyve" to "nl.thehyve"
