--- grails-app/init/nl/hyve/gb/backend/Application.groovy
@@ -1,18 +1,0 @@
-/*
- * Copyright (c) 2019  The Hyve B.V.
- *  This file is distributed under the GNU Affero General Public License
- *  (see accompanying file LICENSE).
- */
-
-package nl.hyve.gb.backend
-
-import grails.boot.GrailsApp
-import grails.boot.config.GrailsAutoConfiguration
-import org.springframework.context.annotation.ComponentScan
-
-@ComponentScan
-class Application extends GrailsAutoConfiguration {
-    static void main(String[] args) {
-        GrailsApp.run(Application, args)
-    }
-}
