--- src/main/groovy/nl/hyve/gb/backend/exception/InvalidArgumentsException.groovy
@@ -1,12 +1,0 @@
-/*
- * Copyright (c) 2019  The Hyve B.V.
- *  This file is distributed under the GNU Affero General Public License
- *  (see accompanying file LICENSE).
- */
-
-package nl.hyve.gb.backend.exception
-
-import groovy.transform.InheritConstructors
-
-@InheritConstructors
-class InvalidArgumentsException extends RuntimeException { }
