+++ src/main/groovy/nl/thehyve/gb/backend/exception/InvalidArgumentsException.groovy
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2019  The Hyve B.V.
+ *  This file is distributed under the GNU Affero General Public License
+ *  (see accompanying file LICENSE).
+ */
+
+package nl.thehyve.gb.backend.exception
+
+import groovy.transform.InheritConstructors
+
+@InheritConstructors
+class InvalidArgumentsException extends RuntimeException { }
