+++ src/main/groovy/nl/thehyve/gb/backend/user/User.groovy
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2019  The Hyve B.V.
+ *  This file is distributed under the GNU Affero General Public License
+ *  (see accompanying file LICENSE).
+ */
+
+package nl.thehyve.gb.backend.user
+
+import groovy.transform.Immutable
+
+@Immutable
+class User {
+    String username
+    String realName
+}
